requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
SQLAlchemy==2.0.30
python-dateutil==2.9.0post0
pyyaml==6.0.1
werkzeug==3.0.3